Implement SignalR communication in test component
The changes include implementing SignalR communication to handle asynchronous operations in the test component. This is done by subscribing to an observable for HTTP and SignalR responses. Also, the new implementation includes dispatching actions for HTTP requests and SignalR responses. The responses are also dispatched as actions to the Redux store for application state management.