Implement loading data from ngrx store
The main focus of this commit is the incorporation of ngrx store into the application to manage state and perform asynchronous actions. This involves creating a DemoItemFacade service which is responsible for loading state from the store and providing observables. Utilized this facade in AppComponent, DemoItemDetailComponent, and DemoItemOverviewComponent. Furthermore, the existing ngrx store setup was refined with comprehensive action handlers and additional selectors.