Refactor UI components and enhance data binding
The code introduces several modifications to improve data binding and structure of the application. Major changes include the introduction of a new 'DemoItemFormComponent' that encapsulates form-related logic, refactoring 'DemoItemAddComponent' to use this new component, and enhancing the 'DemoItemDetailComponent' to display demo item details in a more dynamic way. In addition, the selector 'selectedItem' was added to fetch currently selected item from the state.